Tidy-up Hawk client fixtures.
