+++ conftest.py
@@ -10,6 +10,7 @@ from django.db.models.signals import post_save
 from elasticsearch.helpers.test import get_test_client
 from pytest_django.lazy_django import skip_if_no_django
 
+from datahub.core.test_utils import HawkAPITestClient
 from datahub.dnb_api.utils import format_dnb_company
 from datahub.documents.utils import get_s3_client_for_bucket
 from datahub.metadata.test.factories import SectorFactory
@@ -72,6 +73,12 @@ def api_client():
 
     from rest_framework.test import APIClient
     return APIClient()
+
+
+@pytest.fixture
+def hawk_api_client():
+    """Hawk API client fixture."""
+    yield HawkAPITestClient()
 
 
 class _ReturnValueTracker:
