+++ datahub/search/company/test/test_public_search_view.py
@@ -10,7 +10,6 @@ from rest_framework.reverse import reverse
 from datahub.company.models import Company
 from datahub.company.test.factories import CompanyFactory
 from datahub.core import constants
-from datahub.core.test_utils import HawkAPITestClient
 from datahub.search.company import CompanySearchApp
 
 # Index objects for this search app only
@@ -51,12 +50,6 @@ def setup_data(es_with_collector):
         archived=True,
     )
     es_with_collector.flush_and_refresh()
-
-
-@pytest.fixture
-def hawk_api_client():
-    """Hawk API client fixture."""
-    yield HawkAPITestClient()
 
 
 @pytest.fixture
